requirements.txt
fastapi
uvicorn
langchain
langgraph
crewai
psycopg2-binary
sqlalchemy
python-dotenv
google-generativeai
pydantic
scikit-learn
python-docx
openpyxl
python-pptx
playwright
pyautogui
pywinauto
watchdog
litellm
